requirements.txt
backports.functools-lru-cache==1.5
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.15.0
opencv-python-headless==3.4.2.17
Pillow==5.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
subprocess32==3.5.2
tqdm==4.24.0
